refactor(p1): extract Group B SystemAccess domains — Media/Location/Audio/Notifications/UsageStats
Five sibling factory files, one per capability, following the established pattern:

- MediaAccessTools.kt — createMediaSearchTool(context, deps) preserves the
  pre-flight multi-capability ensureGranted loop (deps.permissionBroker reaches
  through the wrapper for "all" media-type fan-out); queryMedia goes file-private.
- LocationAccessTools.kt — createLocationCurrentTool(context, deps) carries the
  withTimeoutOrNull / suspendCancellableCoroutine path verbatim; latestLocation
  + currentOrLatestLocation file-private. @SuppressLint preserved.
- AudioAccessTools.kt — createAudioRecordOnceTool(context, deps) keeps the
  MediaRecorder lifecycle (prepare/start/delay/stop/release) intact;
  recordAudioOnce(context, durationMillis) file-private.
- NotificationsAccessTools.kt — createNotificationListTool(deps) takes only
  deps because the underlying AmberNotificationListenerService snapshot is
  context-free; queryNotifications stateless.
- UsageStatsAccessTools.kt — createUsageStatsListTool(context, deps); appLabel
  moves here from the coordinator (file-private; sole caller is queryUsageStats).

SystemAccessTools coordinator: 5 tool fields collapse to `by lazy { createXxxTool(...) }`
delegators. SystemAccessTools.kt: 823 → 521 lines (-37% from start of Group B,
1441 → 521 cumulative = -64%).

Side cleanup — 13 imports drop from SystemAccessTools.kt:
  SuppressLint, UsageStatsManager, ContentUris, Location, LocationManager,
  MediaRecorder, CancellationSignal, MediaStore, kotlinx.coroutines.delay,
  suspendCancellableCoroutine, withTimeoutOrNull, AmberNotificationListenerService,
  kotlin.coroutines.resume.

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
